requirements.txt
Django==3.2.15
djangorestframework==3.13.1
django-cors-headers==3.13.0

django-environ==0.9.0

django-ckeditor==6.3.2

boto3==1.24.41
django-storages==1.12.3
Pillow==9.2.0

pyodbc==4.0.34
mssql-django==1.2
psycopg2-binary

argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0

gunicorn==20.1.0
whitenoise==6.2.0

psutil~=5.9.4
openpyxl==3.0.10
pandas==1.5.2
selenium~=4.7.2
requests==2.28.1
pytz==2022.6
emoji==2.2.0
PyPDF2==1.26.0
ipdb
pyvmomi
paramiko==2.9.2

keras==2.13.1rc0
tensorflow==2.13.0rc0
streamlit==1.8.0
streamlit-webrtc==0.36.1
opencv-python-headless==4.5.5.62
# pygame==2.1.2
opencv-python
scipy
